requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
motor==3.3.2
pymongo==4.6.1
python-dotenv==1.0.1
cloudinary==1.39.0
PyPDF2==3.0.1
python-docx==1.1.0
httpx==0.27.0
PyJWT==2.8.0
python-multipart==0.0.9
pydantic[email]==2.6.1
